+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Connectivity table:
   be used when using its skeletons in other widgets (new checkbox: "Original
   color").
 
+- The colors used in the Connectivity Graph Plot are now the ones assigned to the
+  input skeletons, i.e. the ones selected in the Connectiviy Widget. These in
+  turn are either its input skeleton color or the default skeleton color.
+
 
 Miscellaneous: